+++ src/JitterDemo/Demos/Demo20.cs
@@ -31,8 +31,8 @@ public struct CollisionTriangle : ISupportMappable
 
     public void SupportMap(in JVector direction, out JVector result)
     {
-        float min = JVector.Dot(A, direction);
-        float dot = JVector.Dot(B, direction);
+        double min = JVector.Dot(A, direction);
+        double dot = JVector.Dot(B, direction);
 
         result = A;
         if (dot > min)
@@ -50,7 +50,7 @@ public struct CollisionTriangle : ISupportMappable
 
     public void GetCenter(out JVector point)
     {
-        point = (1.0f / 3.0f) * (A + B + C);
+        point = (1.0d / 3.0d) * (A + B + C);
     }
 }
 
@@ -93,7 +93,7 @@ public class CustomCollisionDetection : IBroadPhaseFilter
             ts.C = octree.Vertices[octree.Indices[index].IndexC];
 
             bool hit = NarrowPhase.MPREPA(ts, rbs, rbs.RigidBody!.Orientation, rbs.RigidBody!.Position,
-                out JVector pointA, out JVector pointB, out JVector normal, out float penetration);
+                out JVector pointA, out JVector pointB, out JVector normal, out double penetration);
 
             if (hit)
             {
