+++ src/JitterDemo/Demos/Player/Player.cs
@@ -14,7 +14,7 @@ public class Player
     public RigidBody Body { get; }
     public AngularMotor AngularMovement { get; }
 
-    private readonly float capsuleHalfHeight;
+    private readonly double capsuleHalfHeight;
     private readonly World world;
 
     public Player(World world, JVector position)
@@ -27,7 +27,7 @@ public class Player
         // disable velocity damping
         Body.Damping = (0, 0);
 
-        capsuleHalfHeight = cs.Radius + cs.Length * 0.5f;
+        capsuleHalfHeight = cs.Radius + cs.Length * 0.5d;
 
         this.world = world;
 
@@ -35,8 +35,8 @@ public class Player
         Body.DeactivationTime = TimeSpan.MaxValue;
 
         // Add two arms - to be able to visually tell how the player is orientated
-        var arm1 = new TransformedShape(new BoxShape(0.2f, 0.8f, 0.2f), new JVector(+0.5f, 0.3f, 0));
-        var arm2 = new TransformedShape(new BoxShape(0.2f, 0.8f, 0.2f), new JVector(-0.5f, 0.3f, 0));
+        var arm1 = new TransformedShape(new BoxShape(0.2d, 0.8d, 0.2d), new JVector(+0.5d, 0.3d, 0));
+        var arm2 = new TransformedShape(new BoxShape(0.2d, 0.8d, 0.2d), new JVector(-0.5d, 0.3d, 0));
 
         // Add the shapes without recalculating mass and inertia, we take mass and inertia from the capsule
         // shape we added before.
@@ -48,7 +48,7 @@ public class Player
         ur.Initialize(JVector.UnitY, AngularLimit.Full);
 
         // Add some friction
-        Body.Friction = 0.8f;
+        Body.Friction = 0.8d;
 
         // An angular motor for turning.
         AngularMovement = world.CreateConstraint<AngularMotor>(Body, world.NullBody);
@@ -56,7 +56,7 @@ public class Player
         AngularMovement.MaximumForce = 1000;
     }
 
-    public void SetAngularInput(float rotate)
+    public void SetAngularInput(double rotate)
     {
         AngularMovement.TargetVelocity = rotate;
     }
@@ -149,10 +149,10 @@ public class Player
             if (numContacts == 0) continue;
 
             // divide the result by the number of contact points to get the "center" of the contact
-            hitPoint *= (1.0f / numContacts);
+            hitPoint *= (1.0d / numContacts);
 
             // check if the hitpoint is on the players base
-            if (hitPoint.Y <= -0.8f) return true;
+            if (hitPoint.Y <= -0.8d) return true;
         }
 
         hitPoint = JVector.Zero;
@@ -169,7 +169,7 @@ public class Player
         float delta = lambda - capsuleHalfHeight;
 
         hitPoint = Body.Position - JVector.UnitY * lambda;
-        return (hit && delta < 0.04f && floor != null);
+        return (hit && delta < 0.04d && floor != null);
         */
     }
 
@@ -177,20 +177,20 @@ public class Player
     {
         if (CanJump(out RigidBody? floorBody, out JVector hitPoint))
         {
-            float newYVel = 5.0f;
+            double newYVel = 5.0d;
 
             if (floorBody != null && floorBody != null)
             {
                 newYVel += floorBody.Velocity.Y;
             }
 
-            float deltaVel = Body.Velocity.Y - newYVel;
+            double deltaVel = Body.Velocity.Y - newYVel;
 
             Body.Velocity = new JVector(Body.Velocity.X, newYVel, Body.Velocity.Z);
 
             if (floorBody != null && floorBody.IsStatic)
             {
-                float force = Body.Mass * deltaVel * 100.0f;
+                double force = Body.Mass * deltaVel * 100.0d;
                 floorBody.SetActivationState(true);
                 floorBody.AddForce(JVector.UnitY * force, hitPoint);
             }
@@ -204,20 +204,20 @@ public class Player
             return;
         }
 
-        deltaMove *= 3.0f;
+        deltaMove *= 3.0d;
 
-        float deltaMoveLen = deltaMove.Length();
+        double deltaMoveLen = deltaMove.Length();
 
         JVector bodyVel = Body.Velocity;
         bodyVel.Y = 0;
 
-        float bodyVelLen = bodyVel.Length();
+        double bodyVelLen = bodyVel.Length();
 
-        if (deltaMoveLen > 0.01f)
+        if (deltaMoveLen > 0.01d)
         {
             if (bodyVelLen < 5f)
             {
-                var force = JVector.Transform(deltaMove, Body.Orientation) * 10.0f;
+                var force = JVector.Transform(deltaMove, Body.Orientation) * 10.0d;
 
                 Body.AddForce(force);
